bots in custom games and more bot levels

+++ PVPNetConnect/RiotObjects/Platform/Catalog/Champion/ChampionSkinDTO.cs
@@ -60,7 +60,7 @@ namespace PVPNetConnect.RiotObjects.Platform.Catalog.Champion
         public Int32 WinCountRemaining { get; set; }
 
         [InternalName("purchaseDate")]
-        public Int64 PurchaseDate { get; set; }
+        public Double PurchaseDate { get; set; }
 
         [InternalName("endDate")]
         public Int32 EndDate { get; set; }
